# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Steak API Documentation
  description: Steak is a full-stack platform inspired by Steam and Epic Games, designed for buying,
    downloading, and managing licensed PC games. It supports both players and game publishers.
  version: 1.0.0
servers:
  - url: 'https://api.steak.io.vn/'
  - url: 'http://localhost:8888/'
...
paths:
  /api/v1/user/auth/register:
    post:
      tags:
        - user
      summary: Gửi yêu cầu tạo tài khoản người dùng
      description: Gửi yêu cầu tạo tài khoản người dùng (kiểm tra thông tin và gửi email xác nhận)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml#/components/schemas/UserRegisterRequest"
      responses:
        200:
          description: Thông tin tài khoản hợp lệ, check mail để xác nhận tài khoản
        409:
          description: Thông tin tài khoản đã tồn tại trên hệ thống
          content:
            application/json:
              schema:
                $ref: "./components.yaml#/components/schemas/Error"
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml#/components/schemas/Error"
  /api/v1/user/auth/username-login:
    post:
      tags:
        - user
      summary: Đăng nhập tài khoản
      description: Đăng nhập vào tài khoản bằng username
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/UsernameLoginRequest"

      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/UserLoginResponse"
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/user/auth/email-login:
    post:
      tags:
        - user
      summary: Đăng nhập tài khoản
      description: Đăng nhập vào tài khoản bằng email
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/EmailLoginRequest"

      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/UserLoginResponse"
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/auth/register:
    post:
      tags:
        - dev
      summary: Đăng ký nhà phát hành
      description: Đăng ký nhà phát hành
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/PublisherRegisterRequest"
      responses:
        200:
          description: Thông tin tài khoản hợp lệ, check mail để xác nhận tài khoản
        409:
          description: Thông tin tài khoản đã tồn tại trên hệ thống
          content:
            application/json:
              schema:
                $ref: "./components.yaml#/components/schemas/Error"
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml#/components/schemas/Error"
  /api/v1/dev/auth/username-login:
    post:
      tags:
        - dev
      summary: Đăng nhập tài khoản nhà phát hành
      description: Đăng nhập vào tài khoản
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/UsernameLoginRequest"
      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: publisher123
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/auth/email-login:
    post:
      tags:
        - dev
      summary: Đăng nhập tài khoản nhà phát hành
      description: Đăng nhập vào tài khoản bằng email
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/EmailLoginRequest"
      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: publisher123
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/admin/auth/username-login:
    post:
      tags:
        - admin
      summary: Đăng nhập vào tài khoản quản trị
      description: Đăng nhập vào tải khoản quản trị bằng username
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/UsernameLoginRequest"
      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: admin123
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/admin/auth/email-login:
    post:
      tags:
        - admin
      summary: Đăng nhập vào tài khoản quản trị
      description: Đăng nhập vào tải khoản quản trị bằng email
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/EmailLoginRequest"
      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: admin123@gmail.com
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml#/components/schemas/Error"
...
  /api/v1/dev/project/create:
    post:
      tags:
        - publisher-upload
      summary: Tạo mới project game
      description: Tạo mới một project game cho publisher
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: Tạo project thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectId:
                    type: integer
                    example: 900000000000001
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/project/update:
    post:
      tags:
        - publisher-upload
      summary: Lưu nháp thông tin project game
      description: Lưu nháp thông tin project game (draft)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/SaveProjectRequest"
      responses:
        200:
          description: Lưu nháp thành công
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/project/update-build:
    post:
      tags:
        - publisher-upload
      summary: Cập nhật build cho project game
      description: Cập nhật build cho project game
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/UpdatePreBuildRequest"
      responses:
        200:
          description: Cập nhật build thành công
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/project/publish:
    post:
      tags:
        - publisher-upload
      summary: Đăng game lên hệ thống
      description: Đăng game lên hệ thống (publish)
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Đăng game thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Game published successfully
        400:
          description: Project ID không hợp lệ
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/project/list:
    get:
      tags:
        - publisher-upload
      summary: Lấy danh sách project game của publisher
      description: Lấy danh sách project game của publisher, có thể lọc theo trạng thái, từ khoá, phân trang
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: keyword
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: Danh sách project game
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      type: object
        400:
          description: Tham số phân trang không hợp lệ
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
...